get the data to dashboard and add view and sessions user on pgae

+++ Fish_Manage/Repository/IRepository/IOrderRepository.cs
@@ -4,7 +4,7 @@ namespace Fish_Manage.Repository.IRepository
 {
     public interface IOrderRepository : IRepository<Order>
     {
-        Double GetMoneyPerTerm(int term);
+        decimal GetMoneyPerTerm(int term);
         Task<Order> UpdateAsync(Order entity);
         ApplicationUser UserBuyMost(int term);
     }
